+++ packages/imperative/src/profiles/src/doc/parms/index.ts
@@ -9,7 +9,6 @@
 *
 */
 
-export * from "./IDeleteProfile";
 export * from "./ILoadAllProfiles";
 export * from "./ILoadProfile";
 export * from "./IProfileManager";