+++ packages/imperative/src/profiles/src/doc/response/index.ts
@@ -9,6 +9,5 @@
 *
 */
 
-export * from "./IProfileDeleted";
 export * from "./IProfileLoaded";
 export * from "./IProfileValidated";